feat: moved load initial data into a hook
