+++ src/components/root.tsx
@@ -1,10 +1,10 @@
 import { ReactComponent as IconConnect } from '@/assets/icons/connect.svg';
 import { ReactComponent as IconLogo } from '@/assets/icons/logo.svg';
-import { ItemsContextProvider } from '@powerhousedao/design-system';
 import { useSetAtom } from 'jotai';
 import React, { Suspense, useEffect } from 'react';
 import { Outlet } from 'react-router-dom';
 import { useDropFile } from 'src/hooks';
+import { useLoadInitialData } from 'src/hooks/useLoadInitialData';
 import { isElectron, isMac } from 'src/hooks/utils';
 import { userAtom } from 'src/store';
 import Sidebar from './sidebar';
@@ -12,6 +12,7 @@ import Sidebar from './sidebar';
 const ROOT_FILE_DROP = false;
 
 const Root = () => {
+    useLoadInitialData();
     const ref = React.useRef(null);
 
     const setUser = useSetAtom(userAtom);
@@ -37,44 +38,42 @@ const Root = () => {
     }
 
     return (
-        <ItemsContextProvider>
-            <div className="h-screen">
-                {isElectron && (
-                    <div
-                        className={`h-8 w-full
+        <div className="h-screen">
+            {isElectron && (
+                <div
+                    className={`h-8 w-full
                     ${isMac && 'justify-center'}
                     flex items-center bg-gray-50
                     [-webkit-app-region:drag]`}
-                    >
-                        <IconLogo className="ml-1 mr-0.5 p-1.5" />
-                        <IconConnect className="h-3 w-fit" />
-                    </div>
-                )}
-                <div
-                    className={`flex items-stretch overflow-auto
+                >
+                    <IconLogo className="ml-1 mr-0.5 p-1.5" />
+                    <IconConnect className="h-3 w-fit" />
+                </div>
+            )}
+            <div
+                className={`flex items-stretch overflow-auto
                         ${isElectron ? 'h-app-height' : 'h-screen'}
                         ${isDropTarget ? 'bg-slate-50' : 'bg-white'}
                     `}
-                    {...dropProps}
-                    role="presentation"
-                    tabIndex={0}
-                >
-                    <Suspense>
-                        <Sidebar />
-                        <div className="relative flex-1 overflow-auto">
-                            <Outlet />
-                        </div>
-                        <div
-                            ref={ref}
-                            className={`pointer-events-none fixed inset-0 bg-current
+                {...dropProps}
+                role="presentation"
+                tabIndex={0}
+            >
+                <Suspense>
+                    <Sidebar />
+                    <div className="relative flex-1 overflow-auto">
+                        <Outlet />
+                    </div>
+                    <div
+                        ref={ref}
+                        className={`bg-current pointer-events-none fixed inset-0
                             transition-opacity duration-150 ease-in-out
                             ${isDropTarget ? 'opacity-10' : 'opacity-0'}
                         `}
-                        ></div>
-                    </Suspense>
-                </div>
+                    ></div>
+                </Suspense>
             </div>
-        </ItemsContextProvider>
+        </div>
     );
 };
